requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
Werkzeug==3.0.3
